scripts/requirements.txt
google-genai>=0.3.0
playwright>=1.45.0
python-dotenv>=1.0.0
Pillow>=10.0.0
